fix: wrong url path when switch locale
