+++ src/components/search/SearchBar.tsx
@@ -65,7 +65,7 @@ const CustomSearchBox = ({ currentRefinement, refine }: any) => {
   const handleKeyDown = (e) => {
     if (e.key === "Enter" && currentRefinement.trim()) {
       e.preventDefault();
-      router.push(`/search?q=${encodeURIComponent(currentRefinement.trim())}`);
+      router.push(`search?q=${encodeURIComponent(currentRefinement.trim())}`);
     }
   };
