+++ src/i18n/routing.ts
@@ -1,4 +1,4 @@
-import { createSharedPathnamesNavigation } from "next-intl/navigation";
+import { createNavigation } from "next-intl/navigation";
 import { defineRouting } from "next-intl/routing";
 
 export const routing = defineRouting({
@@ -7,28 +7,7 @@ export const routing = defineRouting({
 
   // Used when no locale matches
   defaultLocale: "en",
-  // pathnames: {
-  //   "/": "/",
-  //   "/about-us": {
-  //     en: "/about-us",
-  //     vi: "/ve-chung-toi",
-  //   },
-  //   "/our-collections": {
-  //     en: "/our-collections",
-  //     vi: "/bo-suu-tap",
-  //   },
-  //   "/blogs": {
-  //     en: "/blogs",
-  //     vi: "/blogs",
-  //   },
-  // },
 });
 
-// export type Pathnames = keyof typeof routing.pathnames;
-// export type Locale = typeof routing.locales[number];
-
-// export const { Link, getPathname, redirect, usePathname, useRouter } =
-//   createLocalizedPathnamesNavigation(routing);
-
 export const { Link, redirect, usePathname, useRouter } =
-  createSharedPathnamesNavigation(routing);
+  createNavigation(routing);
