+++ src/middleware.ts
@@ -5,6 +5,7 @@ export default createMiddleware(routing);
 
 export const config = {
   // Match only internationalized pathnames
-  matcher: ["/", "/(vi|en)/:path*"],
+  // matcher: ["/", "/(vi|en)/:path*"],
+  matcher: "/((?!api|trpc|_next|_vercel|.*\\..*).*)",
   localeDetection: true,
 };
